cleanup
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16419 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzrobots/Tank.h
@@ -27,8 +27,8 @@ struct Tank {
   std::string callsign, team, flag;
   bool paused, alive, frozen, super;
 
-  float position[3], angle;
-  float velocity[3], angularVelocity;
+  double position[3], angle;
+  double velocity[3], angularVelocity;
 };
 
 std::ostream& operator<<(std::ostream& os, const Tank& tank);
